some refactor and moving of run download functions
