+++ robot-server/robot_server/runs/router/file_download_router.py
@@ -1,10 +1,8 @@
 """Router for /runs/{runId}/download file bundle endpoints."""
 
-import json
-from datetime import datetime, timezone
 from pathlib import Path
 from textwrap import dedent
-from typing import Annotated, List, Literal, Optional, Tuple, Union
+from typing import Annotated, List, Literal, Tuple, Union
 
 from fastapi import Depends, status
 from fastapi.responses import FileResponse
@@ -20,7 +18,6 @@ from robot_server.data_files.zip_utils import (
     collect_existing_run_images,
     collect_existing_run_output_csvs,
     create_download_staging_dir,
-    sanitize_filename_component,
     write_zip_for_download,
 )
 from robot_server.errors.error_responses import ErrorBody, ErrorDetails
@@ -28,11 +25,17 @@ from robot_server.persistence.fastapi_dependencies import (
     get_persistence_directory_root,
 )
 from robot_server.protocols.dependencies import get_protocol_store
-from robot_server.protocols.protocol_store import ProtocolNotFoundError, ProtocolStore
-from robot_server.runs.dependencies import get_run_data_manager, get_run_store
+from robot_server.protocols.protocol_store import ProtocolStore
+from robot_server.runs.dependencies import get_run_data_manager
 from robot_server.runs.run_data_manager import RunDataManager
-from robot_server.runs.run_models import RunCommandSummary, RunNotFoundError
-from robot_server.runs.run_store import BadRunResource, RunResource, RunStore
+from robot_server.runs.run_models import RunNotFoundError
+from robot_server.runs.run_download_utils import (
+    build_download_artifact_name,
+    collect_protocol_file,
+    collect_rtp_csv,
+    collect_labware_offsets,
+    collect_run_log,
+)
 
 file_download_router = LightRouter()
 
@@ -73,7 +76,6 @@ class NoDownloadContent(ErrorDetails):
 )
 async def download_run_files(
     runId: str,
-    run_store: Annotated[RunStore, Depends(get_run_store)],
     run_data_manager: Annotated[RunDataManager, Depends(get_run_data_manager)],
     data_files_store: Annotated[DataFilesStore, Depends(get_data_files_store)],
     protocol_store: Annotated[ProtocolStore, Depends(get_protocol_store)],
@@ -92,7 +94,7 @@ async def download_run_files(
         persistence_directory_root: Persistence directory used for download staging.
     """
     try:
-        run = run_store.get(runId)
+        run = run_data_manager.get(runId)
     except RunNotFoundError as e:
         raise RunNotFound(detail=str(e)).as_error(status.HTTP_404_NOT_FOUND) from e
 
@@ -107,16 +109,16 @@ async def download_run_files(
             collect_existing_run_images(
                 runId,
                 data_files_store,
-                archive_prefix=_build_download_artifact_name(
+                archive_prefix=build_download_artifact_name(
                     run=run, protocol_store=protocol_store, suffix="_images"
                 ),
             )
         )
-        protocol_entry = _collect_protocol_file(run=run, protocol_store=protocol_store)
+        protocol_entry = collect_protocol_file(run=run, protocol_store=protocol_store)
         if protocol_entry is not None:
             zip_entries.append(protocol_entry)
 
-        rtp_csv_entry = _collect_rtp_csv(
+        rtp_csv_entry = collect_rtp_csv(
             run_id=runId,
             run_data_manager=run_data_manager,
             data_files_store=data_files_store,
@@ -124,9 +126,8 @@ async def download_run_files(
         if rtp_csv_entry is not None:
             zip_entries.append(rtp_csv_entry)
 
-        run_log_entry = _collect_run_log(
+        run_log_entry = collect_run_log(
             run_id=runId,
-            run=run,
             run_data_manager=run_data_manager,
             protocol_store=protocol_store,
             staging_dir=staging_path,
@@ -134,9 +135,8 @@ async def download_run_files(
         if run_log_entry is not None:
             zip_entries.append(run_log_entry)
 
-        offsets_entry = _collect_labware_offsets(
+        offsets_entry = collect_labware_offsets(
             run_id=runId,
-            run=run,
             run_data_manager=run_data_manager,
             protocol_store=protocol_store,
             staging_dir=staging_path,
@@ -148,7 +148,7 @@ async def download_run_files(
             collect_existing_run_output_csvs(
                 runId,
                 data_files_store,
-                archive_prefix=_build_download_artifact_name(
+                archive_prefix=build_download_artifact_name(
                     run=run, protocol_store=protocol_store, suffix="_output"
                 ),
             )
@@ -176,200 +176,3 @@ async def download_run_files(
         background=BackgroundTask(staging_dir.cleanup),
     )
 
-
-def _collect_protocol_file(
-    run: Union[RunResource, BadRunResource],
-    protocol_store: ProtocolStore,
-) -> Optional[Tuple[Path, str]]:
-    """Return the protocol main file for the zip, or None if unavailable."""
-    if run.protocol_id is None:
-        return None
-
-    try:
-        protocol = protocol_store.get(run.protocol_id)
-    except ProtocolNotFoundError:
-        return None
-
-    main_file = protocol.source.main_file
-    if not main_file.exists() or not main_file.is_file():
-        return None
-
-    name_stem = _protocol_download_name_stem(run=run, protocol_store=protocol_store)
-    return (main_file, f"{name_stem}{main_file.suffix}")
-
-
-def _collect_rtp_csv(
-    run_id: str,
-    run_data_manager: RunDataManager,
-    data_files_store: DataFilesStore,
-) -> Optional[Tuple[Path, str]]:
-    """Return the CSV runtime parameter input file, or None if unavailable."""
-    try:
-        run_record = run_data_manager.get(run_id)
-        for param in run_record.runTimeParameters:
-            if param.type == "csv_file" and param.file is not None:
-                file_info = data_files_store.get(param.file.id)
-                file_path = Path(file_info.path)
-                if file_path.exists() and file_path.is_file():
-                    return (file_path, file_info.name)
-        return None
-    except Exception:
-        return None
-
-
-def _collect_run_log(
-    run_id: str,
-    run: Union[RunResource, BadRunResource],
-    run_data_manager: RunDataManager,
-    protocol_store: ProtocolStore,
-    staging_dir: Path,
-) -> Optional[Tuple[Path, str]]:
-    """Build a run log JSON file in ``staging_dir``, or None if unavailable."""
-    try:
-        run_record = run_data_manager.get(run_id)
-        length_probe = run_data_manager.get_commands_slice(
-            run_id=run_id,
-            cursor=0,
-            length=0,
-            include_fixit_commands=True,
-        )
-        command_slice = run_data_manager.get_commands_slice(
-            run_id=run_id,
-            cursor=0,
-            length=length_probe.total_length,
-            include_fixit_commands=True,
-        )
-        command_summaries = [
-            RunCommandSummary.model_construct(
-                id=c.id,
-                key=c.key,
-                commandType=c.commandType,
-                intent=c.intent,
-                status=c.status,
-                createdAt=c.createdAt,
-                startedAt=c.startedAt,
-                completedAt=c.completedAt,
-                params=c.params,
-                error=c.error,
-                notes=c.notes,
-                failedCommandId=c.failedCommandId,
-                commandAnnotationIds=c.commandAnnotationIds
-                if c.commandAnnotationIds
-                else None,
-            ).model_dump(mode="json", by_alias=True)
-            for c in command_slice.commands
-        ]
-
-        run_details = {
-            "data": run_record.model_dump(mode="json", by_alias=True),
-            "commands": {
-                "data": command_summaries,
-                "meta": {
-                    "cursor": command_slice.cursor,
-                    "totalLength": command_slice.total_length,
-                },
-            },
-        }
-
-        archive_name = _build_run_log_filename(run=run, protocol_store=protocol_store)
-        return _write_staging_json(run_details, archive_name, staging_dir)
-    except Exception:
-        return None
-
-
-def _collect_labware_offsets(
-    run_id: str,
-    run: Union[RunResource, BadRunResource],
-    run_data_manager: RunDataManager,
-    protocol_store: ProtocolStore,
-    staging_dir: Path,
-) -> Optional[Tuple[Path, str]]:
-    """Build labware offsets JSON in ``staging_dir``, or None if unavailable."""
-    try:
-        run_record = run_data_manager.get(run_id)
-        offsets_payload = [
-            offset.model_dump(mode="json", by_alias=True)
-            for offset in run_record.labwareOffsets
-        ]
-        archive_name = _build_labware_offsets_filename(
-            run=run, protocol_store=protocol_store
-        )
-        return _write_staging_json(offsets_payload, archive_name, staging_dir)
-    except Exception:
-        return None
-
-
-def _write_staging_json(
-    payload: object, archive_name: str, staging_dir: Path
-) -> Tuple[Path, str]:
-    """Write JSON payload into the request scratch directory for zip inclusion."""
-    file_path = staging_dir / archive_name
-    file_path.parent.mkdir(parents=True, exist_ok=True)
-    with file_path.open(mode="w", encoding="utf-8") as json_file:
-        json.dump(payload, json_file)
-    return (file_path, archive_name)
-
-
-def _build_download_artifact_name(
-    run: Union[RunResource, BadRunResource],
-    protocol_store: ProtocolStore,
-    *,
-    suffix: str,
-) -> str:
-    """Build `{protocolName}_{ISO-timestamp}{suffix}` for zip entries/directories."""
-    created_at = _format_download_timestamp(run.created_at)
-    name_stem = _protocol_download_name_stem(run=run, protocol_store=protocol_store)
-    return f"{name_stem}_{created_at}{suffix}"
-
-
-def _build_run_log_filename(
-    run: Union[RunResource, BadRunResource],
-    protocol_store: ProtocolStore,
-) -> str:
-    """Build `{protocolName}_{ISO-timestamp}.json`."""
-    return _build_download_artifact_name(
-        run=run, protocol_store=protocol_store, suffix=".json"
-    )
-
-
-def _build_labware_offsets_filename(
-    run: Union[RunResource, BadRunResource],
-    protocol_store: ProtocolStore,
-) -> str:
-    """Build `{protocolName}_{ISO-timestamp}_offsetdata.json`."""
-    return _build_download_artifact_name(
-        run=run, protocol_store=protocol_store, suffix="_offsetdata.json"
-    )
-
-
-def _protocol_download_name_stem(
-    run: Union[RunResource, BadRunResource],
-    protocol_store: ProtocolStore,
-) -> str:
-    """Prefer protocolName metadata, then protocol file name, then ids."""
-    if run.protocol_id is not None:
-        try:
-            protocol = protocol_store.get(run.protocol_id)
-        except ProtocolNotFoundError:
-            protocol = None
-
-        if protocol is not None:
-            protocol_name = protocol.source.metadata.get("protocolName")
-            if protocol_name is None:
-                protocol_name = protocol.source.main_file.stem
-            return sanitize_filename_component(str(protocol_name))
-
-        return sanitize_filename_component(run.protocol_id)
-
-    return sanitize_filename_component(run.run_id)
-
-
-def _format_download_timestamp(created_at: datetime) -> str:
-    """Format a run createdAt timestamp like JS toISOString."""
-    if created_at.tzinfo is None:
-        created_at = created_at.replace(tzinfo=timezone.utc)
-    else:
-        created_at = created_at.astimezone(timezone.utc)
-
-    iso = created_at.isoformat(timespec="milliseconds").replace("+00:00", "Z")
-    return iso.replace(":", "_")
